Implemented Oracle-based play algorithms
Previously the engine only supported playing in a total-information context,
which is not useful for tracking real-workd games.  For the sake of konig-real
parallel implementations based on the new Oracle interface have been added.

Testing so far is minimal, but I wanted to commit 'cos the quantity of changes
was growing substantial.

This is a fairly pervasive change, and hoovered up some other issues along the
way:
- More terminal modes for the additional things required from user
- TerminalOracle to provide Oracle implementation fro konig-real
- Assorted indentation fixes
- Stopped accumulating vectors of announcements and bids for no apparent
  purpose
- Some extra indirection in BiddingSequence implementation
- Added Cards::make_deck
- A ConstrainedCard abstraction to capture the constraint of the fourth king
  when we don't know what it is
- Switched various places storing the called king to use boost::optional,
  rather than storing the pagat when the king was unknown or irrelevant
- Contracts now all expose talon_halves (how many halves of the talon declarer
  gets)