More grappling with called kings abds solo scoring
We now have to pass the called king to ContractAndAnnouncements::score, even
though one was passed to its constructor too.  Furthermore, now initialize the
called_king properly in new PositiveContract::play so that other kings aren't
mistaken for the fourth king during play.

Also, corrected solo scoring.  We now only cancel the kontra when the contract
is off, not made.