More efficient FateAi
Pulled CardInt out from being an implementation detail of Cards to a full-blown
header of its own.  Use integer representation of cards to store Fates in a
vector instead of a map, which is much smaller and faster.  FateAi::start_game
now ten times faster, and other bits of FateAi also better.